Fix an integer overflow when dealing with fixedbvs.
Exhibited by 01_cbmc_Fixedbv23, which couldn't actually generate a
correct representation of -0.5f, instead getting an integer overflow and
some vast number.

+++ solvers/z3/z3_conv.cpp
@@ -425,6 +425,7 @@ std::string stringify(double x)
 std::string
 z3_convt::fixed_point(std::string v, unsigned width)
 {
+  // XXX FIXME: vastly susceptible to integer overflows
   const int precision = 1000000;
   std::string i, f, b, result;
   double integer, fraction, base;
@@ -447,7 +448,7 @@ z3_convt::fixed_point(std::string v, unsigned width)
   if (fraction == 0)
     result = double2string(integer);
   else  {
-    int numerator = (integer*precision + fraction);
+    long int numerator = (integer*precision + fraction);
     result = itos(numerator) + "/" + double2string(precision); 
   }
