+++ solvers/z3/z3_conv.h
@@ -196,7 +196,7 @@ private:
 
   expr2tc bv_get_rec(const Z3_ast bv, const type2tc &type);
 
-  std::string itos(int i);
+  std::string itos(long int i);
   std::string fixed_point(std::string v, unsigned width);
   std::string extract_magnitude(std::string v, unsigned width);
   std::string extract_fraction(std::string v, unsigned width);
